remove & move : Search - ListItem, HomeRecyclerAdapter, HomeViewPager 위치 옮김 - FeedFragment 비활성화
