--- app/src/main/java/com/blackcows/butakaeyak/ui/home/FilterBottomSheet.kt
@@ -1,7 +1,0 @@
-package com.blackcows.butakaeyak.ui.home
-
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-class FilterBottomSheet : BottomSheetDialogFragment() {
-
-}
