+++ app/src/main/java/com/blackcows/butakaeyak/ui/home/adapter/HomeRecyclerAdapter.kt
@@ -1,17 +1,13 @@
-package com.blackcows.butakaeyak.ui.home
+package com.blackcows.butakaeyak.ui.home.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageButton
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.blackcows.butakaeyak.R
-import com.blackcows.butakaeyak.databinding.FragmentFeedListBinding
-import com.blackcows.butakaeyak.databinding.FragmentResultListBinding
+import com.blackcows.butakaeyak.ui.home.data.ListItem
 
 import com.blackcows.butakaeyak.ui.home.placeholder.PlaceholderContent.PlaceholderItem
