+++ app/src/main/java/com/blackcows/butakaeyak/ui/home/adapter/HomeViewPager.kt
@@ -1,12 +1,13 @@
-package com.blackcows.butakaeyak.ui.home
+package com.blackcows.butakaeyak.ui.home.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.blackcows.butakaeyak.ui.home.PillResultFragment
 
 class HomeViewPager(fragment: Fragment): FragmentStateAdapter(fragment) {
     //TODO: update list.
-    private val pages by lazy { listOf(PillResultFragment(), FeedFragment()) }
-    val pageTag = listOf(PillResultFragment.TAB_NAME,FeedFragment.TAB_NAME)
+    private val pages by lazy { listOf(PillResultFragment()) }//, FeedFragment()
+    val pageTag = listOf(PillResultFragment.TAB_NAME)//,FeedFragment.TAB_NAME
 
     override fun getItemCount() = pages.size
     override fun createFragment(position: Int) = pages[position]
